App. js store in components folder user fragment in components
