+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\index.js":"1","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\App.js":"2","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\ButtonPanel.js":"3","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Display.js":"4","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Button.js":"5"},{"size":153,"mtime":1610764964541,"results":"6","hashOfConfig":"7"},{"size":321,"mtime":1610840661048,"results":"8","hashOfConfig":"7"},{"size":1011,"mtime":1610825169610,"results":"9","hashOfConfig":"7"},{"size":275,"mtime":1610840897358,"results":"10","hashOfConfig":"7"},{"size":251,"mtime":1610825169609,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"tcwk4s",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\index.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\App.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\ButtonPanel.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Display.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Button.js",[]]
+[{"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\index.js":"1","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\App.js":"2","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Display.js":"3","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\ButtonPanel.js":"4","C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Button.js":"5"},{"size":164,"mtime":1610903892897,"results":"6","hashOfConfig":"7"},{"size":324,"mtime":1610903838092,"results":"8","hashOfConfig":"7"},{"size":301,"mtime":1610904483896,"results":"9","hashOfConfig":"7"},{"size":1011,"mtime":1610903651619,"results":"10","hashOfConfig":"7"},{"size":251,"mtime":1610903651619,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"tcwk4s",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\index.js",[],["23","24","25","26","27","28","29"],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\App.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Display.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\ButtonPanel.js",[],"C:\\Users\\alex_\\OneDrive\\Escritorio\\Microverse_2\\react-calcu\\src\\components\\Button.js",[],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","replacedBy":"43"},"jsx-a11y/accessible-emoji",[],"lines-around-directive",["44"],"no-spaced-func",["45"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
