--- src/App.css
@@ -1,3 +1,0 @@
-.title {
-  text-align: center;
-}
