--- src/App.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import './App.css';
-import Display from './components/Display';
-import ButtonPanel from './components/ButtonPanel';
-
-function App() {
-  return (
-    <div className="App">
-      <h1 className="title">Calculator</h1>
-      <Display />
-      <ButtonPanel />
-    </div>
-  );
-}
-
-export default App;
