+++ src/components/App.css
@@ -1,0 +1,4 @@
+.title {
+    text-align: center;
+  }
+
